Move template rendering to separate file
